aws/lambda/ec2-testercode/requirements.txt
boto3
requests
aws_requests_auth.aws_auth
